Adding class Array rotation and Array sorting in manner
