+++ array-operations/src/main/java/com/simple/example/array_operations/RotateArrayExample.java
@@ -1,20 +1,18 @@
+/**
+ * 
+ */
+package com.simple.example.array_operations;
+
 import java.util.Arrays;
 
 /**
- * This program 1 -> to sort the given array in sequential order of less than
- * and greater than. Eg. Input = [1,5,4,99] output = [1,5,4,99] 1[4,99,5]
- * 4[5,99] 5[99] 99[]
- * 
- * 
- * 
- * 2 -> to rotate the array using given number input { 1, 4, 5, 9, 10, 12, 12 }
+ * This program to rotate the array using given number input { 1, 4, 5, 9, 10, 12, 12 }
  * rotate by 2 output -> {5,9,10,12,12,1,4}
- * 
  * @author sameer.jadhav
  *
  */
-
-public class ArrayOperation {
+public class RotateArrayExample {
+ 
 	/**
 	 * Main method
 	 * 
@@ -22,45 +20,8 @@ public class ArrayOperation {
 	 */
 	public static void main(String[] args) {
 		int[] arr = { 1, 4, 5, 9, 10, 12, 12 };
-
-	    sortArrayInLessGreatermanner(arr);
-
-		System.out.println("****************************");
-
 		rotateArray(arr, 3);
 	}
-
-	/**
-	 * this method will sort the array in given order
-	 * Algorithm:
-	 * 
-	 * Sort the array then start from 3 element and swap the previous and current value 
-	 * increment by 3 
-	 * 
-	 * if array length is not divisible by 3 and  mod is 2 then swap last and second last value.
-	 * 
-	 * @param a
-	 */
-	public static void sortArrayInLessGreatermanner(int[] a) {
-		int b[] = Arrays.copyOf(a, a.length);
-		Arrays.sort(b);
-		int length = b.length;
-		for (int i = 2; i < length; i += 3) {
-			int temp = b[i];
-			b[i] = b[i - 1];
-			b[i - 1] = temp;
-		}
-		// if some array length can not be divided by 3 in this case
-		if (length % 3 == 2) {
-			int temp = b[length - 1];
-			b[length - 1] = b[length - 2];
-			b[length - 2] = temp;
-		}
-		for (int i : b) {
-			System.out.print(i + " ");
-		}
-	}
-
 	/**
 	 * This method rotate given array with given rotations
 	 * Algorithm:
